\name{attachData-methods}
\docType{methods}
\alias{attachData}
\alias{attachData-methods}
\alias{attachData,TextDocCol,TextDocument-method}
\alias{attachData,TextRepository,TextDocCol-method}
\title{Methods for Function attachData in Package `textmin'}
\description{
Methods for function \code{attachData} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", data = "TextDocument"}{Returns a text
  document collection where the \code{data} text document is added to
  the data slot of the \code{object} text document collection.}

\item{object = "TextRepository", data = "TextDocCol"}{Returns a text
  repository where the \code{data} text document collection is added to
  the data slot of the \code{object} text repository.}
}}
\keyword{methods}
