\name{attach_metadata-methods}
\docType{methods}
\alias{attach_metadata}
\alias{attach_metadata-methods}
\alias{attach_metadata,TextDocCol-method}
\alias{attach_metadata,TextRepository-method}
\title{Methods for Function attach\_metadata in Package `textmin'}
\description{
Methods for function \code{attach_metadata} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", name, metadata}{Returns a text
  document collection where the \code{name}-\code{metadata} tag-value
  pair is added to the metadata slot of the \code{object} text document collection.}

\item{object = "TextRepository", name, metadata}{Returns a text
  repository where the \code{name}-\code{metadata} tag-value
  pair is added to the metadata slot of the \code{object} text repository.}
}}
\keyword{methods}
