\name{c-methods}
\docType{methods}
\alias{c}
\alias{c-methods}
\alias{c,TextDocCol-method}
\title{Methods for Function c in Package `textmin'}
\description{
Methods for function \code{c} in Package `textmin'
}
\section{Methods}{
\describe{

\item{x = "TextDocCol"}{concatenates several text document collections
  to a single one.}
}}
\keyword{methods}
