\name{english_stopwords}
\alias{english_stopwords}
\docType{data}
\title{English stopwords}
\description{
  Common stopwords in the English language.
}
\usage{data(english_stopwords)}
\format{
  A character vector containing 37 stopwords.
}
\source{
  \url{http://www.ranks.nl/tools/stopwords.html}
}
\keyword{datasets}
