\name{filterIDs-methods}
\docType{methods}
\alias{filterIDs}
\alias{filterIDs-methods}
\alias{filterIDs,TextDocument,numeric-method}
\title{Methods for Function filterIDs in Package `textmin'}
\description{
Methods for function \code{filterIDs} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocument", IDs = "numeric"}{returns \code{TRUE} if
  the \code{ID} of \code{object} is in \code{IDs}.}
}}
\keyword{methods}
