\name{filterREUT21578Topics-methods}
\docType{methods}
\alias{filterREUT21578Topics}
\alias{filterREUT21578Topics-methods}
\alias{filterREUT21578Topics,PlainTextDocument,character-method}
\title{Methods for Function filterREUT21578Topics in Package `textmin'}
\description{
Methods for function \code{filterREUT21578Topics} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", topics = "character"}{returns
  \code{TRUE} if the topics stored in the \code{Topics} tag of the
  \code{object}'s local metadata match those in \code{topics}.}
}}
\keyword{methods}
