\name{findHighFreqTerms-methods}
\docType{methods}
\alias{findHighFreqTerms}
\alias{findHighFreqTerms-methods}
\alias{findHighFreqTerms,TermDocMatrix,numeric-method}
\title{Methods for Function findHighFreqTerms in Package `textmin'}
\description{
Methods for function \code{findHighFreqTerms} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TermDocMatrix", freq = "numeric"}{Return those terms
  from \code{object} which occur more or equal often than \code{freq} times.
}}}
\keyword{methods}
