\name{fulltext_search_filter-methods}
\docType{methods}
\alias{fulltext_search_filter}
\alias{fulltext_search_filter-methods}
\alias{fulltext_search_filter,PlainTextDocument,character-method}
\title{Methods for Function fulltext\_search\_filter in Package `textmin'}
\description{
  Methods for function \code{fulltext_search_filter} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument", pattern = "character"}{return
      \code{TRUE} if the regular expression \code{pattern} matches
      \code{object}'s text corpus.}
  }
}
\keyword{methods}
