\name{id.filter-methods}
\docType{methods}
\alias{id.filter}
\alias{id.filter-methods}
\alias{id.filter,TextDocument,numeric-method}
\title{Methods for Function id.filter in Package `textmin'}
\description{
Methods for function \code{id.filter} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TextDocument", IDs = "numeric"}{returns \code{TRUE} if
      the \code{ID} of \code{object} is in \code{IDs}.}
  }
}
\keyword{methods}
