\name{loadFileIntoMem-methods}
\docType{methods}
\alias{loadFileIntoMem}
\alias{loadFileIntoMem-methods}
\alias{loadFileIntoMem,PlainTextDocument-method}
\alias{loadFileIntoMem,XMLTextDocument-method}
\alias{loadFileIntoMem,NewsgroupDocument-method}
\title{Methods for Function loadFileIntoMem in Package `textmin'}
\description{
Methods for function \code{loadFileIntoMem} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}

\item{object = "XMLTextDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}

\item{object = "NewsgroupDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}
}}
\keyword{methods}
