\name{modify_metadata-methods}
\docType{methods}
\alias{modify_metadata}
\alias{modify_metadata-methods}
\alias{modify_metadata,TextDocCol-method}
\alias{modify_metadata,TextRepository-method}
\title{Methods for Function modify\_metadata in Package `textmin'}
\description{
  Methods for function \code{modify_metadata} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TextDocCol", name, metadata}{Returns a text
      document collection with the \code{name} metadata being changed to
      \code{metadata}.}
    \item{object = "TextRepository", name, metadata}{Returns a text
      repository with the \code{name} metadata being changed to
      \code{metadata}.}
  }
}
\keyword{methods}
