\name{remove_metadata-methods}
\docType{methods}
\alias{remove_metadata}
\alias{remove_metadata-methods}
\alias{remove_metadata,TextDocCol-method}
\alias{remove_metadata,TextRepository-method}
\title{Methods for Function remove\_metadata in Package `textmin'}
\description{
  Methods for function \code{remove_metadata} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TextDocCol", name}{Returns a text
      document collection with the \code{name} metadata being removed.}
    \item{object = "TextRepository", name}{Returns a text
      repository with the \code{name} metadata being removed.}
  }
}
\keyword{methods}
