\name{removeStopWords-methods}
\docType{methods}
\alias{removeStopWords}
\alias{removeStopWords-methods}
\alias{removeStopWords,PlainTextDocument,character-method}
\title{Methods for Function removeStopWords in Package `textmin'}
\description{
Methods for function \code{removeStopWords} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", stopwords = "character"}{Removes all
words which occur in \code{stopwords} from \code{object} and return the object.}
}}
\keyword{methods}
