\name{reuters21578.topic.filter-methods}
\docType{methods}
\alias{reuters21578.topic.filter}
\alias{reuters21578.topic.filter-methods}
\alias{reuters21578.topic.filter,PlainTextDocument,character-method}
\title{Methods for Function reuters21578.topic.filter in Package `textmin'}
\description{
Methods for function \code{reuters21578.topic.filter} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument", topics = "character"}{returns
      \code{TRUE} if the topics stored in the \code{Topics} tag of the
      \code{object}'s local metadata match those in \code{topics}.}
  }
}
\keyword{methods}
