\name{set_subscriptable-methods}
\docType{methods}
\alias{set_subscriptable}
\alias{set_subscriptable-methods}
\alias{set_subscriptable,TextDocCol-method}
\title{Methods for Function set\_subscriptable in Package `textmin'}
\description{
Methods for function \code{set_subscriptable} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", name}{Marks the metadata tag \code{name} in
\code{objects}'s metadata as subscriptable. That means that the subset
operator \code{[} takes only relevant elements of the metdata when applied.}
}}
\keyword{methods}
