\name{stemTextDocument-methods}
\docType{methods}
\alias{stemTextDocument}
\alias{stemTextDocument-methods}
\alias{stemTextDocument,PlainTextDocument-method}
\title{Methods for Function stemTextDocument in Package `textmin'}
\description{
Methods for function \code{stemTextDocument} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument"}{Stems the corpus from \code{object}
  with Porter's stemming algorithm.}
}}
\keyword{methods}
