\name{step_next-methods}
\docType{methods}
\alias{step_next}
\alias{step_next-methods}
\alias{step_next,DirSource-method}
\alias{step_next,CSVSource-method}
\alias{step_next,ReutersSource-method}
\title{Methods for Function step\_next in Package `textmin'}
\description{
  Methods for function \code{step_next} in Package `textmin'.
}
\section{Methods}{
  \describe{
    \item{object = "DirSource"}{jumps to the next element in the source.}
    \item{object = "CSVSource"}{jumps to the next element in the source.}
    \item{object = "ReutersSource"}{jumps to the next element in the source.}
  }
}
\keyword{methods}
