\name{strip_whitespace-methods}
\docType{methods}
\alias{strip_whitespace}
\alias{strip_whitespace-methods}
\alias{strip_whitespace,PlainTextDocument-method}
\title{Methods for Function strip\_whitespace in Package `textmin'}
\description{
Methods for function \code{strip_whitespace} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", ...}{Strips any unnecessary
whitespace from \code{object} return the object.}
}}
\keyword{methods}
