\name{[-methods}
\docType{methods}
\alias{[}
\alias{[-methods}
\alias{[,TextDocCol-method}
\title{Methods for Function [ in Package `textmin'}
\description{
Methods for function \code{[} in Package `textmin'
}
\section{Methods}{
\describe{

\item{x = "TextDocCol"}{a custom subset operator which guarantees that a
valid \code{TextDocCol} is returned.}
}}
\keyword{methods}
