\name{tm_index-methods}
\docType{methods}
\alias{tm_index}
\alias{tm_index-methods}
\alias{tm_index,TextDocCol-method}
\title{Methods for Function tm\_index in Package `textmin'}
\description{
Methods for function \code{tm_index} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "TextDocCol", FUN}{Applies \code{FUN} onto each element
      of \code{object}. Note that global metadata is automatically passed to
      \code{FUN} as argument \code{GlobalMetaData}. The function \code{FUN}
      must return a Boolean value.}
  }
}
\seealso{
  \code{\link{tm_filter}}
}
\keyword{methods}
