\name{attach_data-methods}
\docType{methods}
\alias{attach_data}
\alias{attach_data-methods}
\alias{attach_data,TextDocCol,TextDocument-method}
\alias{attach_data,TextRepository,TextDocCol-method}
\title{Methods for Function attach\_data in Package `textmin'}
\description{
Methods for function \code{attach_data} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", data = "TextDocument"}{Returns a text
  document collection where the \code{data} text document is added to
  the data slot of the \code{object} text document collection.}

\item{object = "TextRepository", data = "TextDocCol"}{Returns a text
  repository where the \code{data} text document collection is added to
  the data slot of the \code{object} text repository.}
}}
\keyword{methods}
