\name{attachMetaData-methods}
\docType{methods}
\alias{attachMetaData}
\alias{attachMetaData-methods}
\alias{attachMetaData,TextDocCol-method}
\alias{attachMetaData,TextRepository-method}
\title{Methods for Function attachMetaData in Package `textmin'}
\description{
Methods for function \code{attachMetaData} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", name, metadata}{Returns a text
  document collection where the \code{name}-\code{metadata} tag-value
  pair is added to the metadata slot of the \code{object} text document collection.}

\item{object = "TextRepository", name, metadata}{Returns a text
  repository where the \code{name}-\code{metadata} tag-value
  pair is added to the metadata slot of the \code{object} text repository.}
}}
\keyword{methods}
