\name{fulltext.search.filter-methods}
\docType{methods}
\alias{fulltext.search.filter}
\alias{fulltext.search.filter-methods}
\alias{fulltext.search.filter,PlainTextDocument,character-method}
\title{Methods for Function fulltext.search.filter in Package `textmin'}
\description{
  Methods for function \code{fulltext.search.filter} in Package `textmin'
}
\section{Methods}{
  \describe{
    \item{object = "PlainTextDocument", pattern = "character"}{return
      \code{TRUE} if the regular expression \code{pattern} matches
      \code{object}'s text corpus.}
  }
}
\keyword{methods}
