\name{load_doc-methods}
\docType{methods}
\alias{load_doc}
\alias{load_doc-methods}
\alias{load_doc,PlainTextDocument-method}
\alias{load_doc,XMLTextDocument-method}
\alias{load_doc,NewsgroupDocument-method}
\title{Methods for Function load\_doc in Package `textmin'}
\description{
Methods for function \code{load_doc} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}

\item{object = "XMLTextDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}

\item{object = "NewsgroupDocument"}{Loads the text corpus of
  \code{object} from disk into memory and return the new object.}
}}
\keyword{methods}
