\name{plaintext_parser}
\alias{plaintext_parser}
\title{Parse Text Document}
\description{
  Parse a text document without knowledge about internal structure and
  possible available metadata.
}
\usage{
plaintext_parser(...)
}
\arguments{
  \item{...}{arguments for generator function.}
}
\value{
  A \code{PlainTextDocument} representing \code{file}.
}
\author{Ingo Feinerer}
\keyword{file}
