\name{rcv1_parser}
\alias{rcv1_parser}
\title{Parse Reuters Corpus Volume 1 XML Document}
\description{
  Parse a Reuters Corpus Volume 1 XML document.
}
\usage{
rcv1_parser(...)
}
\arguments{
  \item{...}{arguments for generator function.}
}
\value{
  An \code{XMLTextDocument} representing \code{file}.
}
\author{Ingo Feinerer}
\keyword{file}
