\name{remove_words-methods}
\docType{methods}
\alias{remove_words}
\alias{remove_words-methods}
\alias{remove_words,PlainTextDocument,character-method}
\title{Methods for Function remove\_words in Package `textmin'}
\description{
Methods for function \code{remove_words} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument", stopwords = "character"}{Removes all
words which occur in \code{stopwords} from \code{object} and return the object.}
}}
\keyword{methods}
