\name{reuters21578xml_parser}
\alias{reuters21578xml_parser}
\title{Parse Reuters21578 XML Document}
\description{
  Parse a Reuters21578 XML document.
}
\usage{
reuters21578xml_parser(...)
}
\arguments{
  \item{...}{arguments for generator function.}
}
\value{
  An \code{XMLTextDocument} representing \code{file}.
}
\author{Ingo Feinerer}
\keyword{file}
