\name{setSubscriptable-methods}
\docType{methods}
\alias{setSubscriptable}
\alias{setSubscriptable-methods}
\alias{setSubscriptable,TextDocCol-method}
\title{Methods for Function setSubscriptable in Package `textmin'}
\description{
Methods for function \code{setSubscriptable} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "TextDocCol", name}{Marks the metadata tag \code{name} in
\code{objects}'s metadata as subscriptable. That means that the subset
operator \code{[} takes only relevant elements of the metdata when applied.}
}}
\keyword{methods}
