\name{stem_doc-methods}
\docType{methods}
\alias{stem_doc}
\alias{stem_doc-methods}
\alias{stem_doc,PlainTextDocument-method}
\title{Methods for Function stem\_doc in Package `textmin'}
\description{
Methods for function \code{stem_doc} in Package `textmin'
}
\section{Methods}{
\describe{

\item{object = "PlainTextDocument"}{Stems the corpus from \code{object}
  with Porter's stemming algorithm.}
}}
\keyword{methods}
